Redux error 수정
