+++ src/app/(content)/layout.tsx
@@ -1,28 +1,20 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
 import '../globals.css'
-import { Provider } from 'react-redux'
-import store from "../store";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
+import ButtonTest from "../_component/ButtonTeset"
+import ReduxProvider from "../_component/ReduxProvider"
 
 export default function RootLayout({
   children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+}: {
+  children: React.ReactNode
+}) {
   return (
     <html lang="en">
-      <body className={inter.className}>
-        <Provider store={store} >
+      <body>
+        <ReduxProvider>
+          <ButtonTest />
           {children}
-        </Provider>
+        </ReduxProvider>
       </body>
     </html>
-  );
+  )
 }
