+++ src/app/(home)/_component/Main.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export default function Home() {
+    return (
+        <main>
+            <h1>Home Page</h1>
+            <h1>Home Page</h1>
+            <h1>Home Page</h1>
+            <h1>Home Page</h1>
+            <h1>Home Page</h1>
+            <h1>Home Page</h1>
+        </main>
+    );
+}
