+++ src/app/(home)/layout.tsx
@@ -1,9 +1,8 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import '../globals.css'
-import ButtonTest from "./_component/ButtonTeset";
-import { Provider } from "react-redux";
-import store from "../store";
+import ButtonTest from "../_component/ButtonTeset";
+import ReduxProvider from "../_component/ReduxProvider";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -20,10 +19,10 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <Provider store={store}>
+        <ReduxProvider>
           <ButtonTest />
           {children}
-        </Provider>
+        </ReduxProvider>
       </body>
     </html>
   );
