+++ src/app/(home)/page.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
+import Main from './_component/Main';
 
 export default function Home() {
   return (
     <>
-      <h1>Home Page</h1>
+      <Main />
     </>
   );
 }
