+++ src/app/_component/ReduxProvider.tsx
@@ -1,0 +1,14 @@
+'use client'
+import React from "react";
+import { Provider } from "react-redux";
+import store from "../store";
+
+type Props = {
+    children: React.ReactNode;
+}
+
+const ReduxProvider: React.FC<Props> = ({ children }) => {
+    return <Provider store={store}>{children}</Provider>;
+};
+
+export default ReduxProvider;
